src/main/java/ru/bazhanov/utils/StringUtils.java
package ru.bazhanov.utils;

import java.util.Collections;
import java.util.LinkedHashMap;
import java.util.Map;
import java.util.stream.Collectors;

public class StringUtils {

    public static Map<Character,Long> accountingForCharactersInString(String currentString){
        return currentString.toLowerCase().chars()
                .mapToObj(ch -> (char)ch)
                .collect(Collectors.groupingBy(ch -> ch, Collectors.counting()))
                .entrySet()
                .stream()
                .sorted(Collections.reverseOrder(Map.Entry.comparingByValue()))
                .collect(Collectors.toMap(
                        Map.Entry::getKey,
                        Map.Entry::getValue,
                        (a, b) -> { throw new AssertionError(); },
                        LinkedHashMap::new
                ));
    }
}
